[backend] refactor: JPA 호환성을 위해 모든 ID 타입을 long으로 수정

+++ backend/src/main/java/com/example/demo/controller/MemberController.java
@@ -74,7 +74,7 @@ public class MemberController {
 
     // 사용자가 멤버인 팀 리스트 요청
     @GetMapping("/members/userTeamList/{userId}")
-    public HashMap<String, Object> requestUserTeamList(@PathVariable(required = true) int userId) {
+    public HashMap<String, Object> requestUserTeamList(@PathVariable(required = true) long userId) {
         List<TeamRequest> data = memberService.requestUserTeamList(userId);
 
         HashMap<String, Object> result = new HashMap<>();
