+++ backend/src/main/java/com/example/demo/entity/TeamEntity.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class TeamEntity {
     private long id;
     private String team_name;
-    private int creator_id;
+    private long creator_id;
 }
